backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
sqlalchemy==2.0.12
alembic==1.10.4
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
email-validator==2.0.0.post2
python-multipart==0.0.6
requests==2.30.0
python-dotenv==1.0.0
wechatpy==1.8.18
pyjwt==2.7.0
starlette==0.26.1
typing-extensions==4.5.0
aiofiles==23.1.0
qrcode==7.4.2
pillow==9.5.0
pytz==2023.3
tenacity==8.2.2